requirements.txt
rapidfuzz==3.9.6
python-dotenv==1.0.1
supabase==2.6.0